# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Xray Panel API
  version: 0.1.0
  description: >
    Personal multi-VPS Xray management panel. Design-first contract shared by the
    real Go backend and the Phase-1 mock frontend (MSW). No billing, traffic
    accounting, tenancy or referrals — single admin, personal use.
servers:
  - url: /
    description: Same-origin (Caddy proxies /api and /sub to the backend)
...
security:
  - bearerAuth: []
paths:
...
  /api/servers:
    get:
      tags: [Servers]
      summary: List servers
      responses:
        '200':
          description: Servers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Server'
    post:
      tags: [Servers]
      summary: Create a server
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServerInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Server'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/servers/{id}:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    get:
      tags: [Servers]
      summary: Get a server
      responses:
        '200':
          description: Server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Server'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Servers]
      summary: Update a server
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServerInput'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Server'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Servers]
      summary: Delete a server
      responses:
        '204':
          description: Deleted
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
  responses:
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
        code:
          type: string
        details:
          type: object
          additionalProperties: true
    Engine:
      type: string
      enum: [xray, hysteria]
      description: Target engine an inbound belongs to.
...
    ServerStatus:
      type: string
      enum: [online, offline, unknown, error]
    ProvisionStatus:
      type: string
      enum: [pending, installing, installed, failed]
      description: Engine installation (provisioning) state of the node, SPEC §5.1.
    EngineStatus:
      type: object
      required: [engine, running]
      properties:
        engine:
          $ref: '#/components/schemas/Engine'
        running:
          type: boolean
        version:
          type: string
        service_name:
          type: string
    Server:
      type: object
      required: [id, name, host, ssh_port, ssh_user, ssh_auth_method, status]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        host:
          type: string
        ssh_port:
          type: integer
        ssh_user:
          type: string
        ssh_auth_method:
          type: string
          enum: [password, key]
        xray_config_path:
          type: string
        xray_service_name:
          type: string
        hysteria_config_path:
          type: string
        hysteria_service_name:
          type: string
        ip:
          type: string
        geo_country:
          type: string
        geo_city:
          type: string
        geo_asn:
          type: string
        status:
          $ref: '#/components/schemas/ServerStatus'
        provision_status:
          $ref: '#/components/schemas/ProvisionStatus'
        provision_error:
          type: string
          nullable: true
        engines:
          type: array
          items:
            $ref: '#/components/schemas/EngineStatus'
        inbound_count:
          type: integer
        last_check_at:
          type: string
          format: date-time
          nullable: true
        last_sync_at:
          type: string
          format: date-time
          nullable: true
        last_sync_error:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ServerInput:
      type: object
      required: [name, host, ssh_port, ssh_user, ssh_auth_method]
      properties:
        name:
          type: string
        host:
          type: string
        ssh_port:
          type: integer
          default: 22
        ssh_user:
          type: string
        ssh_auth_method:
          type: string
          enum: [password, key]
        ssh_secret:
          type: string
          format: password
          description: SSH password or private key (write-only; encrypted at rest).
        ssh_passphrase:
          type: string
          format: password
          description: Passphrase for the private key (write-only).
        xray_config_path:
          type: string
          default: /usr/local/etc/xray/config.json
        xray_service_name:
          type: string
          default: xray
        hysteria_config_path:
          type: string
          default: /etc/sing-box/config.json
        hysteria_service_name:
          type: string
          default: sing-box
        wipe_existing:
          type: boolean
          default: false
          description: >
            One-time provision option (not stored). If true, the panel removes
            competing proxy/VPN stacks and panels on the node before installing
            the engines, freeing conflicting ports and memory. OS/SSH untouched.
...